fix: small fixes for the tests to run more correctly
